new file:   .Draw.py.swp 	new file:   .Generate.py.swp 	new file:   .MathHelper.py.swp 	new file:   .color.py.swp 	new file:   .intersection_convex_polygon.py.swp 	new file:   .main.py.swp 	new file:   .mutual_arrangement_segment_polygon.py.swp 	modified:   Draw.py 	modified:   Generate.py 	modified:   LinesRelative.py 	modified:   MathHelper.py 	modified:   Triangle.py 	modified:   Triangulation_Delone.py 	modified:   __pycache__/Draw.cpython-37.pyc 	modified:   __pycache__/Generate.cpython-37.pyc 	modified:   __pycache__/IsInSimple.cpython-37.pyc 	modified:   __pycache__/LinesRelative.cpython-37.pyc 	modified:   __pycache__/MathHelper.cpython-37.pyc 	modified:   __pycache__/Point.cpython-37.pyc 	modified:   __pycache__/color.cpython-37.pyc 	new file:   __pycache__/intersection_convex_polygon.cpython-37.pyc 	modified:   __pycache__/mutual_arrangement_segment_polygon.cpython-37.pyc 	modified:   color.py 	new file:   intersection_convex_polygon.py 	modified:   main.py 	modified:   mutual_arrangement_segment_polygon.py

+++ MathHelper.py
@@ -268,7 +268,7 @@ def param_t_for_line(start, point, end):
 def point_from_param_t(a, b, t):
     assert len(a) == 2
     assert len(b) == 2
-    return [a[0] * (1 - t) + b[0] * t, a[1] * (1 - t) + b[1] * t]
+    return [int(a[0] * (1 - t) + b[0] * t), int(a[1] * (1 - t) + b[1] * t)]
 
 def sorted_by_x(points):
     """
@@ -322,26 +322,4 @@ def is_delonay_condition(p1, p2, p3, p4):
                 abs((p2[0] - p4[0]) * (p3[1] - p4[1]) - 
                     (p3[0] - p4[0]) * (p2[1] - p4[1])) * cA            
                 ) >= 0
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
